server/requirements.txt
fastapi
uvicorn
aiogram
aiohttp
aiohttp-socks
python-dotenv
